fix: toast doesn't close even when transaction receipt is there (#524)
The receipt we sent to check was always undefined because we were
accessing a field that didn't exist. We were also retrying after the
transaction was reverted, because we were retrying on any errors. Both
are fixed now

## Summary by Sourcery

Fix toast not closing and retrying logic when minting tokens.

Bug Fixes:
- Fixed an issue where the toast notification would not close even after
a transaction receipt was received.
- Fixed an issue where token minting would retry even after the
transaction was reverted.

Enhancements:
- Improved error handling for token minting transactions.