fix: create stablecoin form validation triggered on opening the form (#523)
On opening the form, the validation was getting triggered and I was
getting errors without having touched any fields. We were manually
triggering the form validation when the component was rendered. Now:

1. Don't trigger react hook form validation manually
2. To check if we can go to the next step, use zod schema validation on
the fields of the form step

Something like `isFieldValid = !field.invalid && field.isDirty` works
when a default value is not provided, but when a default value is
provided `isDirty` is `false` so `isFieldValid` is always `false` even
though there is a valid value. So using the zod schema covers all the
cases here

## Summary by Sourcery

Remove manual form validation and use Zod schema validation to determine
if a form step is valid.

Bug Fixes:
- Fix form validation being triggered on opening the form.

Enhancements:
- Use Zod schema validation for form steps.

---------

Co-authored-by: Roderik van der Veer <roderik@settlemint.com>