feat: Implement caching for asset data fetching (#562)
The changes in this commit introduce caching for the asset data fetching
functions in the admin section of the dapp. The main changes are:

- Wrap the asset data fetching functions (`getEquities()`,
`getCryptocurrencies()`, `getStableCoins()`, and `getBonds()`) with the
`unstable_cache()` function from the `next/cache` module.
- Set the caching options, including a revalidation time of 10 seconds
and tags for each asset type.
- This change will improve the performance of the admin section by
reducing the number of unnecessary GraphQL requests and providing a
better user experience.

## Summary by Sourcery

Enhancements:
- Cache asset data fetched via GraphQL queries to improve performance.