feat: Add ISIN field to create token form
The changes in this commit add an ISIN (International Securities Identification Number) field to the create token form. This field is required and allows users to specify the ISIN for the token they are creating. The ISIN is a unique identifier for a security and is used to track and identify financial instruments.

The changes include:

- Adding the `isin` field to the `CreateTokenSchema` and making it a required field.
- Updating the `CreateTokenMutation` to include the `isin` field in the mutation variables.
- Adding the ISIN input field to the `TokenBasics` component and marking it as required.
- Updating the `createTokenAction` to include the `isin` field in the mutation variables.

These changes will improve the token creation process by allowing users to specify the ISIN for the token, which is an important piece of information for financial instruments.