feat(tokens): Allow optional ISIN field in token creation (#593)
The changes made in this commit allow the ISIN field to be optional when
creating a new token. Previously, the ISIN field was required, but this
has been updated to make it optional. This change provides more
flexibility for token creators, as some tokens may not have an ISIN
associated with them.

The key changes are:

- In the `create-token-action.ts` file, the `isin` field is set to an
empty string if no value is provided.
- In the `1-token-basics.tsx` file, the `TextInput` component for the
ISIN field has been updated to remove the `showRequired` prop, making
the field optional.
- In the `create-token-form-schema.ts` file, the `isin` field in the
`BaseTokenFields` schema has been updated to be optional.

## Summary by Sourcery

New Features:
- Allow token creation without an ISIN.