feat: add trivy scan for codestudio docker image (#605)
This change adds a new step to the CI workflow that runs the Trivy
vulnerability scanner on the `ghcr.io/settlemint/codestudio-starterkit-
asset-tokenization` Docker image. This is done to ensure that the
codestudio image is free of known vulnerabilities before it is deployed.
The scan is only run for non-PR builds.

Additionally, the existing Trivy scan step has been renamed to be more
specific to the custom deployment Docker image.