feat: add all fields, actions on user management section (#807)
## Summary by Sourcery

Add last activity tracking for users and update user KYC status. Display
the last activity in the admin panel and add an action to update the KYC
status. Improve the users table by adding sorting and filtering options.
Update the asset balance fragment to include the paused status for more
asset types. Update the data table row actions component to accept a
render function for children and add a close function to the render
props. Update the data table component to accept an initial sorting prop
and use it to set the initial sorting state.

New Features:
- Display user's last activity on the platform.
- Add an action to update a user's KYC status in the admin panel.
- Add sorting to the users table in the admin panel.
- Add filtering by status and KYC status to the users table in the admin
panel.
- Display the status of assets in the portfolio and asset details pages
using a pill component.
- Update the change role action to use the DetailUser type instead of
the ListUser type.
- Update the asset balance fragment to include the paused status for
bonds, funds, and equities.
- Update the data table row actions component to accept a render
function for children and add a close function to the render props.
- Update the data table component to accept an initial sorting prop and
use it to set the initial sorting state.
- Update the data table component to use the initial sorting state when
the component mounts.