feat: track collateral stats (#808)
## Summary by Sourcery

Add tracking and display of collateral statistics for stablecoins. This
includes displaying the collateral ratio, expiration date, and validity
duration on the admin details page. Update the subgraph to track these
statistics and make them available to the dapp.

New Features:
- Display collateral statistics, including collateral ratio, expiration
date, and validity duration for stablecoins on the admin details page.
- Track and display the collateral ratio for stablecoins in the subgraph
and update it whenever the collateral or total supply changes.
- Add fields for amount of holders, last collateral update, liveness,
and creator to the subgraph data for stablecoins.
- Add and format the display of collateral proof expiration and validity
on the stablecoin details page.
- Add formatting functions for durations to the date library.
- Include ISIN in the stablecoin details page only if it exists.
- Fetch and display the creator of the stablecoin on the details page if
available from the subgraph