feat: redact pin codes in the dev log (#838)
## Summary by Sourcery

This pull request introduces redaction of the `pincode` field in the dev
logs for enhanced security. It also fixes a bug in the BlockUserForm
that was causing the blocked state to be inverted, and updates the
BlockUserForm and BlockButton components to use `currentlyBlocked`
instead of `blocked` for clarity.

New Features:
- Introduces redaction of the `pincode` field in the dev logs for
enhanced security.

Bug Fixes:
- Fixes a bug in the BlockUserForm that was causing the blocked state to
be inverted.

Enhancements:
- Updates the BlockUserForm and BlockButton components to use
`currentlyBlocked` instead of `blocked` for clarity.