fix: do not allow spaces and numbers in symbols (#843)
## Summary by Sourcery

This pull request fixes a bug where spaces and numbers were allowed in
the symbol field of asset forms. It introduces a new `textOnly` prop to
the `AssetFormInput` component, which, when set to true, enforces that
the input only contains letters. This change is applied to the symbol
field in the create forms for bonds, cryptocurrencies, equities, funds,
and stablecoins.

Bug Fixes:
- Restricted the characters allowed in the symbol field of asset forms
to only letters.

Enhancements:
- Added a `textOnly` prop to the `AssetFormInput` component to enforce
text-only input.