feat: detail pages for portfolio (#858)
## Summary by Sourcery

This pull request introduces detailed pages for each asset type in the
user's portfolio, providing comprehensive information and statistics. It
also refactors the data fetching logic and improves performance by
prefetching data on the server side.

New Features:
- Introduces detail pages for different asset types (Stablecoin, Equity,
Bond, Cryptocurrency) within the portfolio, each displaying specific
information and statistics related to the asset.
- Adds a new route `/portfolio/my-assets` to display the user's assets
in a table.

Enhancements:
- Updates the asset table to include a link to the detail page for each
asset.
- Refactors the data fetching logic for MyAssets to be more flexible and
reusable, allowing it to be used in both the dashboard and the detail
pages.
- Improves data fetching by prefetching data on the server side using
`prefetchQuery` and `dehydrate` from `@tanstack/react-query`.

Chores:
- Removes the 'My Contacts' link from the portfolio sidebar.