feat: revoke all (#866)
https://github.com/user-attachments/assets/99359b49-f12e-4449-85f1-6b31eb024c90

## Summary by Sourcery

This pull request introduces a 'Revoke All' feature to the stablecoin
admin panel. This feature enables administrators to remove all
permissions for a user on a specific asset through a new 'Revoke All'
button and associated form.

New Features:
- Introduces a 'Revoke All' feature in the admin panel for stablecoins,
allowing administrators to remove all permissions for a specific user on
a given asset.

Enhancements:
- Adds a 'Revoke All' button to the token permissions table, providing a
user-friendly way to trigger the new functionality.
- Implements a multi-step form using AssetForm, including a summary step
that displays the current roles and requires PIN code confirmation for
security.

---------

Co-authored-by: Silke Van den Broeck <silke@settlemint.com>