feat: withdraw excess underlying asset (#888)
## Summary by Sourcery

This pull request introduces a new feature that allows administrators to
withdraw excess underlying assets from a bond. It adds a new UI
component for initiating the withdrawal, including specifying a
recipient address and confirming the transaction with a PIN code. It
also enhances the asset details grid to display the underlying asset
address as a clickable EvmAddress component.

New Features:
- Introduces the ability for administrators to withdraw excess
underlying assets from a bond.
- Adds a new UI component for initiating the withdrawal of excess
underlying assets, including specifying a recipient address and
confirming the transaction with a PIN code.

Enhancements:
- Displays the underlying asset address as a clickable EvmAddress
component in the asset details grid.

Co-authored-by: Silke Van den Broeck <silke@settlemint.com>