fix: permission table (#953)
## Summary by Sourcery

Fixes an issue where the column definitions for the DataTable component
were not being correctly parameterized, and updates the type of
totalSupply in AssetActivityFragmentSchema from bigint to bigDecimal.

Bug Fixes:
- Fixes the DataTable component to correctly handle parameterized column
definitions, ensuring that the address is properly passed to the columns
function.
- Updates the type of totalSupply in AssetActivityFragmentSchema from
bigint to bigDecimal to align with the actual data type.