fix: cleanup usage of private assets in the UI (#988)
## Summary by Sourcery

Removes the 'private asset' functionality from the UI, including the
switch in the creation forms and the display in tables. Adds form
validation error messages to the form component and makes several input
fields required.

Enhancements:
- Removes the 'private asset' functionality from the UI.
- Adds form validation error messages to the form component.
- Makes several input fields required in the create forms for bonds,
funds, equities, and stablecoins, as well as in the mint, burn, freeze,
and update collateral forms.