feat: account / balance subgraph improvements (#1039)
## Summary by Sourcery

Adds fields to the Account and AssetBalance entities to track active
balances and total balances, and updates the counts when assets are
transferred, paused, or unpaused.

Enhancements:
- Adds fields to the Account entity to track the number of active
balances, the exact total balance, and the total balance in decimal
format.
- Adds a field to the AssetBalance entity to track the last activity
timestamp.
- Updates the Account and AssetBalance entities when assets are
transferred, paused, or unpaused to keep the active balances and total
balances up to date.