feat: my assets / user holdings improvements (#1048)
## Summary by Sourcery

Improves the display of user assets and holdings by adding an asset
status component and updating the columns in various tables to display
the asset status and balance with the token symbol.

Enhancements:
- Replaces the blocked/unblocked icons with an AssetStatusPill component
to display the asset status.
- Updates the columns in the user assets table to display the asset
status and balance with the token symbol.
- Updates the columns in the user holdings table to display the asset
status and balance with the token symbol.
- Refactors the formatAssetStatus utility to handle both AssetBalance
and Asset types.
- Removes the icons object from the columns files as it is no longer
needed.
- Updates the user table to use a badge for the banned status and
improves the display of the role and KYC status.