fix: dedicated portal api to display transactions timeline graph (#1093)
## Summary by Sourcery

Fixes the transaction timeline graph on the dedicated portal API to
display the correct data for the last month and last year. It introduces
new functions to generate timeline data for one month and one year, and
modifies the existing transaction history query to map the data
correctly.

Enhancements:
- Improves the transaction history display by adding functions to
generate timeline data for one month and one year.
- Updates the transaction history query to map the data correctly for
the chart.
- Adds a formatChartDate function to format the date based on the
granularity of the chart.
- Adds a isRawData prop to the TransactionsHistory component to handle
pre-formatted data.