fix: transfer form and portfolio improvements (#1191)
## Summary by Sourcery

Improves the user experience and fixes visibility issues in the ATK
demo. It enhances the transfer form by adding a user search feature and
allows manual entry of wallet addresses. It also displays the user's
balance for assets on the details page.

New Features:
- Adds a user search feature to the transfer form, allowing users to
search for recipients instead of manually entering wallet addresses.
- Displays the user's balance for assets on the details page, providing
more information to the user about their holdings.
- Adds the ability to switch between manual entry and user search in the
recipients step of the transfer form.
- Adds a conditional check for balance data to prevent errors when the
balance is not available or the user is not logged in.
- Adds a default value of empty array to allowlist and blocklist in
AssetUsersFragmentSchema to prevent errors when the lists are empty.

---------

Co-authored-by: Silke Van den Broeck <silke@settlemint.com>