feat: start migration to typebox to unify the schemas between app and api (#1216)
## Summary by Sourcery

Migrate cryptocurrency data validation to TypeBox schemas for improved
type safety and consistency between the app and API layers. This
includes defining schemas for on-chain and off-chain cryptocurrency
data, as well as calculated fields. The existing Zod schemas are
removed, and the code is updated to use the new TypeBox schemas for data
parsing and validation.

Enhancements:
- Replace Zod schemas with TypeBox schemas for cryptocurrency data
validation.
- Improve type safety and consistency between the app and API layers.
- Introduce calculated fields for cryptocurrency data, such as
concentration.
- Refactor cryptocurrency list and detail queries to use the new TypeBox
schemas.
- Add new TypeBox validation utilities.
- Add new calculated fields utility.
- Update currency selection to include GBP and CHF