fix: duplicated transfer forms, recent assets infinite re-render (#1246)
## Summary by Sourcery

Fixes an issue where transfer forms were duplicated and prevents
infinite re-renders of recent assets by ensuring the user address is
converted to lowercase when comparing it to the holder's account ID and
initializing recent assets with an empty array.

Bug Fixes:
- Fixes duplicated transfer forms.
- Prevents infinite re-render of recent assets by converting the user
address to lowercase when comparing it to the holder's account ID.
- Initializes recent assets with an empty array to prevent infinite
re-renders of recent assets.
- Fixes an issue where the assetType was not being passed correctly to
the transferAssetFunction.
- Fixes an issue where the transferAssetFunction was not being called
correctly from the transfer action.